geotiff tests: consolidate GPU codec cluster (Sub-PR C)
Cluster 14 of long-tail epic #2424. Fold 11 GPU codec test files into
a new xrspatial/geotiff/tests/gpu/test_codec.py:

- test_nvcomp_batch_compress_batched_1712.py
- test_nvcomp_batch_upload_p3.py
- test_nvcomp_decompress_cumsum_offsets_1950.py
- test_nvcomp_from_device_bufs_single_alloc_1659.py
- test_nvjpeg_encode_stream_sync_2212.py
- test_nvjpeg2k_single_alloc_2107.py
- test_jpeg_gpu_1549.py
- test_lerc_valid_mask_gpu.py
- test_predictor2_big_endian_gpu_1517.py
- test_predictor3_int_dtype_gpu_1933.py
- test_gpu_jpeg_interop_reject_issue_D_1845.py

Tests-only. The shared requires_gpu marker from _helpers/markers.py
replaces the per-file _gpu_available / _gpu_only helpers. Module-level
helpers carry the source issue number suffix to keep sibling sections
collision-free. 71 tests collected, matching the baseline sum.

The release-gate row for the jpeg codec and the unit/test_predictor
docstring now cite gpu/test_codec.py.

Refs #2438.
